Move clocks to clk module
