+++ src/sys/clk/mod.rs
@@ -1,8 +1,12 @@
 mod cmos;
+mod boot;
+mod epoch;
 mod rtc;
 mod sleep;
 mod timer;
 
+pub use boot::{uptime, Uptime}; // TODO: Rename to boot_time
+pub use epoch::{realtime, Realtime}; // TODO: Rename to epoch_time
 pub use cmos::CMOS;
 pub use rtc::RTC;
 pub use sleep::{sleep, nanowait, halt};
@@ -14,6 +18,20 @@ pub use timer::{
 use rtc::{Interrupt, Register, RTC_CENTURY};
 use timer::{rdtsc, CLOCKS_PER_NANOSECOND};
 
+use crate::api::clock::DATE_TIME_ZONE;
+
+use time::{Duration, OffsetDateTime};
+
 pub fn init() {
     timer::init();
+}
+
+pub fn log_rtc() {
+    let s = realtime();
+    let ns = Duration::nanoseconds(
+        libm::floor(1e9 * (s - libm::floor(s))) as i64
+    );
+    let dt = OffsetDateTime::from_unix_timestamp(s as i64) + ns;
+    let rtc = dt.format(DATE_TIME_ZONE);
+    log!("RTC {}", rtc);
 }
