+++ src/sys/net/socket/tcp.rs
@@ -56,10 +56,10 @@ impl TcpSocket {
     pub fn connect(&mut self, addr: IpAddress, port: u16) -> Result<(), ()> {
         let mut connecting = false;
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 let mut sockets = SOCKETS.lock();
@@ -119,10 +119,10 @@ impl TcpSocket {
 
     pub fn accept(&mut self) -> Result<IpAddress, ()> {
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 let mut sockets = SOCKETS.lock();
@@ -147,12 +147,12 @@ impl TcpSocket {
 impl FileIO for TcpSocket {
     fn read(&mut self, buf: &mut [u8]) -> Result<usize, ()> {
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         let mut bytes = 0;
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             let mut sockets = SOCKETS.lock();
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 iface.poll(sys::net::time(), device, &mut sockets);
@@ -184,12 +184,12 @@ impl FileIO for TcpSocket {
 
     fn write(&mut self, buf: &[u8]) -> Result<usize, ()> {
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         let mut sent = false;
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             let mut sockets = SOCKETS.lock();
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 iface.poll(sys::net::time(), device, &mut sockets);
