+++ src/sys/net/socket/udp.rs
@@ -60,10 +60,10 @@ impl UdpSocket {
 
     pub fn connect(&mut self, addr: IpAddress, port: u16) -> Result<(), ()> {
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 let mut sockets = SOCKETS.lock();
@@ -98,12 +98,12 @@ impl UdpSocket {
 impl FileIO for UdpSocket {
     fn read(&mut self, buf: &mut [u8]) -> Result<usize, ()> {
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             let bytes;
             let mut sockets = SOCKETS.lock();
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 iface.poll(sys::net::time(), device, &mut sockets);
@@ -132,12 +132,12 @@ impl FileIO for UdpSocket {
 
     fn write(&mut self, buf: &[u8]) -> Result<usize, ()> {
         let timeout = 5.0;
-        let started = sys::clock::realtime();
+        let started = sys::clk::realtime();
         let mut sent = false;
         if let Some((ref mut iface, ref mut device)) = *sys::net::NET.lock() {
             let mut sockets = SOCKETS.lock();
             loop {
-                if sys::clock::realtime() - started > timeout {
+                if sys::clk::realtime() - started > timeout {
                     return Err(());
                 }
                 iface.poll(sys::net::time(), device, &mut sockets);
